src/transaction/views/delete_transaction_api_view.py
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework.permissions import IsAuthenticated
from django.shortcuts import get_object_or_404

from transaction.models import Transaction


class DeleteTransactionAPIView(APIView):
    permission_classes = [IsAuthenticated]

    def delete(self, request, transaction_id):
        transaction = get_object_or_404(Transaction, id=transaction_id, user=request.user)
        transaction.delete()
        return Response({'success': True})
