pyproject.toml
[project]
name = "capyfin"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = "==3.12.*"

[dependency-groups]
prod = [
    "django==5.1.6",
    "django-cors-headers==4.7.0",
    "djangorestframework==3.15.2",
    "drf-social-oauth2==3.1.0",
    "gunicorn==23.0.0",
    "psycopg2-binary==2.9.10",
    "pydantic-settings==2.7.1",
    "requests>=2.32.3",
]
dev = [
    "ipython>=8.32.0",
    "ruff==0.8.6",
    "uv==0.5.30",
]

[tool.uv]
default-groups = ["dev", "prod"]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 120
indent-width = 4

target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

docstring-code-format = false

docstring-code-line-length = "dynamic"
